refactor(mcp): move eval-related UI components to eval package #371
- Relocate McpMessageLogPanel, RequestDetailPanel, ResponseDetailPanel, and ParameterDisplay to ui.eval package
- Update imports in affected files
- Clean up unused imports and comments
