+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/ui/eval/McpMessageLogPanel.kt
@@ -1,4 +1,4 @@
-package cc.unitmesh.devti.mcp.ui
+package cc.unitmesh.devti.mcp.ui.eval
 
 import cc.unitmesh.devti.mcp.ui.model.McpMessage
 import cc.unitmesh.devti.mcp.ui.model.MessageType
@@ -32,7 +32,6 @@ class McpMessageLogPanel : JPanel(BorderLayout()) {
     private val detailPanel = JPanel(detailCardLayout)
 
     private val requestDetailPanel = RequestDetailPanel()
-    
     private val responseDetailPanel = ResponseDetailPanel()
     
     companion object {
